ProjectilePool is no longer a Node. Switched OverwriteX/YMovement to MomentumPreservation and BurstMomentumPreservation, adjust values of moves accourdingly. Increased final AUSpecial hitbox size. Moved more variables to be properties. Added backgrounds. Load a texture on the map base and it will automatically be created. Fixed loading 8 characters not actually loading 8 characters. Merged all debug labels into one. Changed and renamed a lot of the PublicData methods. Added TryAdd extension method to dictionaries. Made frame counts a long. Changed GetCooldown() to FinalCooldown for attacks and attack parts. Activating a hitbox will now automatically reset its frame count and run Init. Hitboxes and Hurtboxes now display transparent. Moved collision debug toggles to hurtbox and character collisions so it only updates when needed. Changed more try-catch on dictionary access to checks. Camera will now also follow (0,0), and the camera rect is now limited to prevent zooming out of it. Camera zoom is now linearly interpolated. Made managers not actually run an update if they have no elements. Made managers make an array copy instead of a list copy of their dictionaries. Removed redundant cleanup code in AudioManager. Removed InputDisplayLabel.

+++ Scripts/AttackScripts/AttackPart/AttackPartTransitionManager.cs
@@ -9,7 +9,7 @@ public class AttackPartTransitionManager
 	public void Add(AttackPartTransition apt) => Transitions.Add(apt);
 	public void Clear() => Transitions.Clear();
 	
-	public string NextAttackPart(StateTagsManager tagsManager, int currentFrames) =>
+	public string NextAttackPart(StateTagsManager tagsManager, long currentFrames) =>
 		Transitions
 		.FirstOrDefault(
 			t => t.MatchesTag(tagsManager, currentFrames),
